backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.5
cssselect2==0.8.0
daphne==4.2.1
Django==5.1.6
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.1
et_xmlfile==2.0.0
fonttools==4.59.0
gunicorn==23.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jalali_core==1.0.0
jdatetime==5.2.0
lxml==6.0.0
MarkupSafe==3.0.2
mysqlclient==2.2.7
numpy==2.3.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.1
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.9.0
pyOpenSSL==25.1.0
pypdf==5.8.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
reportlab==4.2.5
requests==2.32.3
sendgrid==6.12.4
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.1
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.9.0
xhtml2pdf==0.2.17
zope.interface==7.2
zopfli==0.2.3.post1
